Move JavaScript implementations into core
