+++ core/src/js/DuktapeContext.cpp
